refactor: code reafactored
separate helper function
combine document/index logic code
